+++ CI/SCRIPTS/__init__.py
@@ -1,0 +1,3 @@
+from shapely import speedups
+
+speedups.disable()