+++ CI/SCRIPTS/test_rasters.py
@@ -60,7 +60,7 @@ def test_rasters():
     # Create tmp file
     # VRT needs to be build on te same disk
     with tempfile.TemporaryDirectory() as tmp_dir:
-        tmp_dir = rasters_path().joinpath("OUTPUT_XARRAY")
+        # tmp_dir = rasters_path().joinpath("OUTPUT_XARRAY")
         os.makedirs(tmp_dir, exist_ok=True)
 
         # Get Extent
@@ -239,14 +239,8 @@ def test_rasters():
             vect_xda = rasters.vectorize(raster_path)
             vect_val = rasters.vectorize(raster_path, values=val)
             vect_val_diss = rasters.vectorize(raster_path, values=val, dissolve=True)
-            vect_val_diss.to_file(
-                os.path.join(tmp_dir, "vect_val_diss.geojson"), driver="GeoJSON"
-            )
             vect_val_disc = rasters.vectorize(
                 raster_path, values=[1, 255], keep_values=False
-            )
-            vect_xda.to_file(
-                os.path.join(tmp_dir, "test_vector.geojson"), driver="GeoJSON"
             )
             ci.assert_geom_equal(vect_xda, vect_truth)
             ci.assert_geom_equal(vect_val_diss, diss_truth)